Remove cyclical dependency from productPrices
