+++ support-frontend/assets/helpers/productPrice/promotions.tsx
@@ -10,7 +10,6 @@ import type {
 	ProductPrice,
 	ProductPrices,
 } from 'helpers/productPrice/productPrices';
-// eslint-disable-next-line import/no-cycle -- these are quite tricky to unpick so we should come back to this
 import { getProductPrice, isNumeric } from 'helpers/productPrice/productPrices';
 import type { SubscriptionProduct } from 'helpers/productPrice/subscriptions';
 import type { Option } from 'helpers/types/option';
@@ -165,6 +164,31 @@ function promotionHTML(
 	);
 }
 
+function finalPrice(
+	productPrices: ProductPrices,
+	country: IsoCountry,
+	billingPeriod: BillingPeriod,
+	fulfilmentOption: FulfilmentOptions = NoFulfilmentOptions,
+	productOption: ProductOptions = NoProductOptions,
+): ProductPrice {
+	return applyDiscount(
+		getProductPrice(
+			productPrices,
+			country,
+			billingPeriod,
+			fulfilmentOption,
+			productOption,
+		),
+		getPromotion(
+			productPrices,
+			country,
+			billingPeriod,
+			fulfilmentOption,
+			productOption,
+		),
+	);
+}
+
 export {
 	getPromotion,
 	getAppliedPromo,
@@ -174,4 +198,5 @@ export {
 	promoQueryParam,
 	getPromotionCopy,
 	promotionHTML,
+	finalPrice,
 };
